refactor(music-player): rename `position_card` folder to `bottom_bar`
