+++ lib/music_player/bottom_bar/button_panel.dart
