+++ lib/music_player/bottom_bar/current_song_panel.dart
