+++ lib/music_player/bottom_bar/highlighted_section_slider_track_shape.dart
