+++ lib/music_player/bottom_bar/position_slider.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:musbx/music_player/looper/looper.dart';
-import 'package:musbx/music_player/position_card/highlighted_section_slider_track_shape.dart';
+import 'package:musbx/music_player/bottom_bar/highlighted_section_slider_track_shape.dart';
 import 'package:musbx/music_player/music_player.dart';
 
 class PositionSlider extends StatelessWidget {
