+++ lib/music_player/music_player_page.dart
@@ -12,9 +12,9 @@ import 'package:musbx/music_player/pick_song_button/components/spacer.dart';
 import 'package:musbx/music_player/pick_song_button/components/search_youtube_button.dart';
 import 'package:musbx/music_player/pick_song_button/components/upload_file_button.dart';
 import 'package:musbx/music_player/slowdowner/slowdowner_card.dart';
-import 'package:musbx/music_player/position_card/button_panel.dart';
-import 'package:musbx/music_player/position_card/current_song_panel.dart';
-import 'package:musbx/music_player/position_card/position_slider.dart';
+import 'package:musbx/music_player/bottom_bar/button_panel.dart';
+import 'package:musbx/music_player/bottom_bar/current_song_panel.dart';
+import 'package:musbx/music_player/bottom_bar/position_slider.dart';
 import 'package:musbx/music_player/song.dart';
 import 'package:musbx/music_player/song_source.dart';
 import 'package:musbx/music_player/pick_song_button/speed_dial.dart';
